erfwong_sup (erfwong v0.1.0)

守护进程

Functions

init(_)

Whenever a supervisor is started using supervisor:start_link/[2,3], this function is called by the new process to find out about restart strategy, maximum restart frequency and child specifications.

start_link()

开启守护进程